refactor: move test to their own folder
