+++ test/entities/email-validation.spec.ts
@@ -1,4 +1,4 @@
-import { Email } from './email'
+import { Email } from '../../src/entities/email'
 
 describe('Email validation', () => {
   test('should not accept null strings', () => {
