+++ test/entities/user.spec.ts
@@ -1,4 +1,4 @@
-import { User } from './user'
+import { User } from '../../src/entities/user'
 
 describe('User class entity', () => {
   test('should not create user with invalid email address', () => {
