+++ test/usecases/register-user-on-maling-list/repository/in-memory-user-repository.spec.ts
@@ -1,4 +1,4 @@
-import { UserData } from '../../../entities/user-data'
+import { UserData } from '../../../../src/entities/user-data'
 import { InMemoryUserRepository } from './in-memory-user-repository'
 
 describe('In Memory User Repository', () => {
