+++ test/usecases/register-user-on-maling-list/repository/in-memory-user-repository.ts
@@ -1,5 +1,5 @@
-import { UserRepository } from '../ports/user-repository'
-import { UserData } from '../../../entities/user-data'
+import { UserRepository } from '../../../../src/usecases/register-user-on-maling-list/ports/user-repository'
+import { UserData } from '../../../../src/entities/user-data'
 
 export class InMemoryUserRepository implements UserRepository {
   private repository: UserData[]
